learntrac-api/requirements.txt
# Core FastAPI
fastapi>=0.104.0
uvicorn[standard]>=0.24.0

# Database
psycopg[binary,pool]>=3.1.0
asyncpg>=0.29.0
sqlalchemy[asyncio]>=2.0.0
alembic>=1.12.0

# Cache & Storage
redis>=5.0.0
neo4j>=5.14.0

# Configuration & Validation
pydantic>=2.5.0
pydantic-settings>=2.1.0

# HTTP & Auth
httpx>=0.25.0
aiohttp>=3.9.0
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.0
python-multipart>=0.0.6

# AWS
boto3>=1.29.0
aioboto3>=12.0.0

# AI/ML
openai>=1.3.0
numpy>=1.26.0
pandas>=2.1.0
sentence-transformers>=2.2.0

# Additional dependencies
cryptography>=41.0.0
python-dotenv>=1.0.0